ADD: new constraint UniqueEntityProperty EDIT: refactored UniqueEntity constraint with parent
