+++ composer.json
@@ -19,7 +19,7 @@
         "sort-packages": true
     },
     "require": {
-        "php": "^7.2",
+        "php": ">=7.2",
         "symfony/http-kernel": "^v4.4.13|^v5.1.5",
         "symfony/dependency-injection": "^v4.4.13|^v5.1.5",
         "symfony/validator": "^v4.4.13|^v5.1.5",
